requirements.txt
pyserial
pymongo
paho-mqtt

